\name{StanfordExampleData}
\alias{RDI108\_n.gpr}
\alias{StanfordDCV1.7complete.txt}
\title{Example of MEEBO GPR files and associated doping control information.}

\description{
  RDI108\_n.gpr is an example of hybridization on MEEBO arrays, using a
  doping-control mixture from SFGF.
  
  StanfordDCV1.7complete.txt is an example of spike-type file from SFGF.
}

\source{These data were provided by members of the Brown lab (Stanford) and the
  Stanford Shared Genomics Core Facility.}

\keyword{datasets}




